# cmd/compile: hoist some loop invariants