… src/cmd/compile/internal/ssa/likelyadjust.go
… src/cmd/compile/internal/ssa/shortcircuit.go